read yotas from file
